Fixed name for asset import window ui.
