+++ editor/CMakeLists.txt
@@ -38,6 +38,7 @@ add_library(${PROJECT_NAME} STATIC
         src/core/ui/imgui/imgui_window_renderer.cpp
         src/core/ui/views/opened_project/menu_bar_ui.cpp
         src/core/ui/views/opened_project/windows/asset_browser_ui.cpp
+        src/core/ui/views/opened_project/windows/asset_import_ui.cpp
         src/core/ui/views/opened_project/windows/details_ui.cpp
         src/core/ui/views/opened_project/windows/scene_outliner_ui.cpp
         src/core/ui/views/opened_project/windows/scene_view_ui.cpp
