+++ editor/src/core/ui/views/opened_project/windows/asset_browser_ui.h
@@ -1,9 +1,1 @@
 #pragma once
-
-#include "../../../imgui/imgui_helper.h"
-
-namespace OpenedProjectUI {
-namespace Windows {
-ImGuiHelper::Window GetAssetBrowserWindow();
-} // namespace Windows
-} // namespace OpenedProjectUI
