+++ editor/src/core/ui/views/opened_project/windows/asset_import_ui.h
@@ -1,1 +1,9 @@
 #pragma once
+
+#include "../../../imgui/imgui_helper.h"
+
+namespace OpenedProjectUI {
+namespace Windows {
+ImGuiHelper::Window GetAssetImportWindow();
+} // namespace Windows
+} // namespace OpenedProjectUI
